backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
slowapi==0.1.9
pydantic[email]
email-validator

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.11

# Supabase
supabase==2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Stripe
stripe==7.8.0

# AI/ML
scikit-learn==1.3.2
sentence-transformers>=2.2.0
lightgbm==4.1.0
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Translation
deep-translator==1.11.4

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1

# Testing
pytest>=7.0.0